remove old scorers
