--- packages/core/src/scores/scorer.test.ts
@@ -1,214 +1,0 @@
-import { describe, it, expect, beforeEach } from 'vitest';
-import {
-  AsyncFunctionBasedScorerBuilders,
-  FunctionBasedScorerBuilders,
-  MixedScorerBuilders,
-  PromptBasedScorerBuilders,
-} from './scorer.test-utils';
-
-const createTestData = () => ({
-  inputText: 'test input',
-  outputText: 'test output',
-  get userInput() {
-    return [{ role: 'user', content: this.inputText }];
-  },
-  get agentOutput() {
-    return { role: 'assistant', text: this.outputText };
-  },
-  get scoringInput() {
-    return { input: this.userInput, output: this.agentOutput };
-  },
-});
-
-describe('createScorer', () => {
-  let testData: ReturnType<typeof createTestData>;
-
-  beforeEach(() => {
-    testData = createTestData();
-  });
-
-  describe('Steps as functions scorer', () => {
-    it('should create a basic scorer with functions', async () => {
-      const scorer = FunctionBasedScorerBuilders.basic;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('should create a scorer with reason', async () => {
-      const scorer = FunctionBasedScorerBuilders.withReason;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('should create a scorer with preprocess and reason', async () => {
-      const scorer = FunctionBasedScorerBuilders.withPreprocessAndReason;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('should create a scorer with preprocess and analyze', async () => {
-      const scorer = FunctionBasedScorerBuilders.withPreprocessAndAnalyze;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('should create a scorer with preprocess only', async () => {
-      const scorer = FunctionBasedScorerBuilders.withPreprocess;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('should create a scorer with preprocess, analyze, and reason', async () => {
-      const scorer = FunctionBasedScorerBuilders.withPreprocessAndAnalyzeAndReason;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('should create a scorer with analyze only', async () => {
-      const scorer = FunctionBasedScorerBuilders.withAnalyze;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('should create a scorer with analyze and reason', async () => {
-      const scorer = FunctionBasedScorerBuilders.withAnalyzeAndReason;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-  });
-
-  describe('Steps as prompt objects scorer', () => {
-    it('with analyze prompt object', async () => {
-      const scorer = PromptBasedScorerBuilders.withAnalyze;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with preprocess and analyze prompt object', async () => {
-      const scorer = PromptBasedScorerBuilders.withPreprocessAndAnalyze;
-
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with analyze and reason prompt object', async () => {
-      const scorer = PromptBasedScorerBuilders.withAnalyzeAndReason;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(typeof result.reason).toBe('string');
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with generate score as prompt object', async () => {
-      const scorer = PromptBasedScorerBuilders.withGenerateScoreAsPromptObject;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with all steps', async () => {
-      const scorer = PromptBasedScorerBuilders.withAllSteps;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-  });
-
-  describe('Mixed scorer', () => {
-    it('with preprocess function and analyze prompt object', async () => {
-      const scorer = MixedScorerBuilders.withPreprocessFunctionAnalyzePrompt;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with preprocess prompt and analyze function', async () => {
-      const scorer = MixedScorerBuilders.withPreprocessPromptAnalyzeFunction;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with reason function and analyze prompt', async () => {
-      const scorer = MixedScorerBuilders.withReasonFunctionAnalyzePrompt;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with reason prompt and analyze function', async () => {
-      const scorer = MixedScorerBuilders.withReasonPromptAnalyzeFunction;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-  });
-
-  describe('Async scorer', () => {
-    it('with basic', async () => {
-      const scorer = AsyncFunctionBasedScorerBuilders.basic;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with preprocess', async () => {
-      const scorer = AsyncFunctionBasedScorerBuilders.withPreprocess;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with preprocess function and analyze as prompt object', async () => {
-      const scorer = AsyncFunctionBasedScorerBuilders.withPreprocessFunctionAndAnalyzePromptObject;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with preprocess prompt object and analyze function', async () => {
-      const scorer = AsyncFunctionBasedScorerBuilders.withPreprocessPromptObjectAndAnalyzeFunction;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with async createPrompt in preprocess', async () => {
-      const scorer = AsyncFunctionBasedScorerBuilders.withAsyncCreatePromptInPreprocess;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with async createPrompt in analyze', async () => {
-      const scorer = AsyncFunctionBasedScorerBuilders.withAsyncCreatePromptInAnalyze;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with async createPrompt in generateScore', async () => {
-      const scorer = AsyncFunctionBasedScorerBuilders.withAsyncCreatePromptInGenerateScore;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-
-    it('with async createPrompt in generateReason', async () => {
-      const scorer = AsyncFunctionBasedScorerBuilders.withAsyncCreatePromptInGenerateReason;
-      const result = await scorer.run(testData.scoringInput);
-
-      expect(result).toMatchSnapshot();
-    });
-  });
-});
